Make ListContacts a controlled component
Revert the definition of ListContacts to a class component. Give this
component a state that tracks a ‘query’ string. Bind this ‘query’
string to the search input element and update the state of this
controlled component on a change to the input element. The ‘query’
string should be updated to whatever the user inputs into the input
element which will in turn update the value of the input